rename - #4: user 도메인 패키지 경로 수정
